--- a/plan/common_plans.go
+++ b/plan/common_plans.go
@@ -188,7 +188,7 @@ func (e *Execute) rebuildRange(p Plan) error {
 		if err != nil {
 			return errors.Trace(err)
 		}
-	default:
+	case PhysicalPlan:
 		var err error
 		for _, child := range x.Children() {
 			err = e.rebuildRange(child)
@@ -422,10 +422,10 @@ func (e *Explain) prepareTaskDot(p PhysicalPlan, taskTp string, buffer *bytes.Bu
 		return
 	}
 
-	var copTasks []Plan
+	var copTasks []PhysicalPlan
 	var pipelines []string
 
-	for planQueue := []Plan{p}; len(planQueue) > 0; planQueue = planQueue[1:] {
+	for planQueue := []PhysicalPlan{p}; len(planQueue) > 0; planQueue = planQueue[1:] {
 		curPlan := planQueue[0]
 		switch copPlan := curPlan.(type) {
 		case *PhysicalTableReader: