--- a/plan/stringer.go
+++ b/plan/stringer.go
@@ -25,12 +25,23 @@ func ToString(p Plan) string {
 }
 
 func toString(in Plan, strs []string, idxs []int) ([]string, []int) {
-	if len(in.Children()) > 1 {
-		idxs = append(idxs, len(strs))
-	}
+	switch x := in.(type) {
+	case LogicalPlan:
+		if len(x.Children()) > 1 {
+			idxs = append(idxs, len(strs))
+		}
 
-	for _, c := range in.Children() {
-		strs, idxs = toString(c, strs, idxs)
+		for _, c := range x.Children() {
+			strs, idxs = toString(c, strs, idxs)
+		}
+	case PhysicalPlan:
+		if len(x.Children()) > 1 {
+			idxs = append(idxs, len(strs))
+		}
+
+		for _, c := range x.Children() {
+			strs, idxs = toString(c, strs, idxs)
+		}
 	}
 
 	var str string