store LogicalPlan and PhysicalPlan directly, avoid type assertion.