@hanfei1991 2018-01-11
/run-sqllogic-test